fixed linter & version compatibility problems

+++ include/coord/pool.h
@@ -10,7 +10,7 @@ typedef struct {
 
 int pool_init();
 void pool_free();
-int pool_redirect(int fd);
+long pool_redirect(int fd);
 int pool_submit(Command *cmd);
 void pool_broadcast(Command *cmd);
 void pool_show();
